refactor: restructure go project
